generate cart with JS and organized files
