+++ shopping-cart/index.html
@@ -32,27 +32,11 @@
                 </thead>
                 <tbody>
                   <tr>    
-                    <td>Zombie</td>
-                    <td>$12</td>
-                    <td>2</td>
-                    <td>$24</td>
-                  </tr>
-                  <tr>
-                    <td>Singapore Sling</td>
-                    <td>$9</td>
-                    <td>3</td>
-                    <td>$27</td>
-                  </tr>
-                  <tr>
-                    <td>Piña colada</td>
-                    <td>$8</td>
-                    <td>1</td>
-                    <td>$8</td>
                   </tr>
                 </tbody>
                 <tfoot>
                     <tr>
-                        <td>Total $60</td>
+                        <td class="total">Total $60</td>
                     </tr> 
                 </tfoot>
             </table>
@@ -63,7 +47,7 @@
         <div class="copy gutter-right">&copy; 2020 Alchemy Code Lab</div>
     </footer>
 
-    <script type="module" src="products.js"></script>
+    <script type="module" src="shopping-cart.js"></script>
 </body>
 
 </html>
